parse to double instead of int

+++ Quadratic.java
@@ -1,8 +1,8 @@
 class Quadratic {
     public static void main(String[] args) {
-      int a = Integer.parseInt(args[0]);
-      int b = Integer.parseInt(args[1]);
-      int c = Integer.parseInt(args[2]);
+      double a = Double.parseDouble(args[0]);
+      double b = Double.parseDouble(args[1]);
+      double c = Double.parseDouble(args[2]);
 
       if (args.length < 3) {
         System.out.println("provide sufficient arguements.");
